+++ bubble/src/Predicates/PassLibrary.cpp
@@ -49,12 +49,6 @@ static PassPtr gate_translation_pass(
   return ptr;
 }
 
-const PassPtr &SynthesiseIBM() {
-  static const PassPtr pp(gate_translation_pass(
-      Transform::synthesise_IBM(),
-      {OpType::U1, OpType::U2, OpType::U3, OpType::CX}, true, "SynthesiseIBM"));
-  return pp;
-}
 const PassPtr &SynthesiseTket() {
   static const PassPtr pp(gate_translation_pass(
       Transform::synthesise_tket(), {OpType::tk1, OpType::CX}, true,
@@ -90,13 +84,6 @@ const PassPtr &RebaseCirq() {
 const PassPtr &RebaseTket() {
   static const PassPtr pp(gate_translation_pass(
       Transform::rebase_tket(), {OpType::CX, OpType::tk1}, true, "RebaseTket"));
-  return pp;
-}
-
-const PassPtr &RebaseIBM() {
-  static const PassPtr pp(gate_translation_pass(
-      Transform::rebase_IBM(), {OpType::CX, OpType::U3, OpType::U2, OpType::U1},
-      true, "RebaseIBM"));
   return pp;
 }
 
@@ -341,16 +328,16 @@ const PassPtr &DecomposeBoxes() {
   return pp;
 }
 
-const PassPtr &USquashIBM() {
+const PassPtr &SquashTK1() {
   static const PassPtr pp([]() {
-    Transform t = Transform::u_squash_IBM();
+    Transform t = Transform::squash_1qb_to_tk1();
     PredicatePtrMap s_ps;
     PredicateClassGuarantees g_postcons{
         {typeid(GateSetPredicate), Guarantee::Clear}};
     PostConditions postcon{s_ps, g_postcons, Guarantee::Preserve};
     // record pass config
     nlohmann::json j;
-    j["name"] = "USquashIBM";
+    j["name"] = "SquashTK1";
     return std::make_shared<StandardPass>(s_ps, t, postcon, j);
   }());
   return pp;