+++ bubble/tests/test_Combinators.cpp
@@ -46,7 +46,7 @@ SCENARIO("List sequencing") {
     const StateVector s0 = tket_sim::get_statevector(circ);
     std::vector<Transform> seq = {
         Transform::optimise_via_PhaseGadget(), Transform::clifford_simp(),
-        Transform::rebase_IBM()};
+        Transform::rebase_tket()};
     Transform::sequence(seq).apply(circ);
     REQUIRE(circ.count_gates(OpType::CX) == 8);
     REQUIRE(circ.count_gates(OpType::V) == 0);